requirements.txt
aldryn-bootstrap3==1.1.1
aldryn-style==1.1.1
cmsplugin-filer==1.1.2
dj-database-url==0.4.1
Django==1.8
django-appconf==1.0.2
django-classy-tags==0.7.2
django-cms==3.2.5
django-durationfield==0.5.2
django-filer==1.2.4
django-formtools==1.0
django-ipware==1.1.5
django-mptt==0.8.4
django-polymorphic==0.8.1
django-reversion==1.8.7
django-sekizai==0.9.0
Django-Select2==4.3.2
django-treebeard==4.0.1
djangocms-admin-style==1.2.3
djangocms-attributes-field==0.1.1
djangocms-column==1.6.0
djangocms-forms==0.2.1
djangocms-googlemap==0.5.1
djangocms-inherit==0.2.1
djangocms-installer==0.8.11
djangocms-link==1.8.2
djangocms-style==1.7.0
djangocms-text-ckeditor==2.9.4
easy-thumbnails==2.3
hashids==1.1.0
html5lib==0.9999999
jsonfield==1.0.3
Pillow==3.3.0
polls==0.0.4
pytz==2016.6.1
requests==2.10.0
six==1.10.0
tablib==0.11.2
tzlocal==1.2.2
Unidecode==0.4.19
virtualenv==15.0.2
virtualenvwrapper-win==1.2.1
